+++ src/helper/init_program.py
@@ -1,10 +1,10 @@
 import logging
 from dotenv import load_dotenv
-from src.definitions import DIR_OUTPUT
-from logs.config.logging import logs_config
-from src.modules.helper.misc import delete_dir_contents
-from src.modules.helper.time import utc_now
-from src.modules.init.pandas_opts import pandas_opts
+from definitions import DIR_OUTPUT
+from logs_setup.config.logging import logs_config
+from helper.misc import delete_dir_contents
+from helper.time import utc_now
+from helper.pandas_opts import pandas_opts
 
 
 def init_program():
