+++ src/run.py
@@ -1,5 +1,5 @@
-from src.modules.init.init_program import init_program
-
+from helper.init_program import init_program
+from definitions import PATH_LOGS_CONFIG
 def main():
 
     # init
@@ -7,6 +7,7 @@ def main():
 
     # actions to perform...
     print("hello world!")
+    print(PATH_LOGS_CONFIG)
